refactor: OutputView 분리 및 DOM 조작 로직 개선
